# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TransitOps fleet operations API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  fleet_manager: 'fleet_manager',
  driver: 'driver',
  safety_officer: 'safety_officer',
  financial_analyst: 'financial_analyst',
} as const;

export interface User {
  id: number;
  clerkUserId: string;
  email: string;
  name: string;
  role: UserRole | null;
  createdAt: string;
}

export interface UserRoleInput {
  role: UserRole;
}

export type VehicleStatus = typeof VehicleStatus[keyof typeof VehicleStatus];


export const VehicleStatus = {
  Available: 'Available',
  On_Trip: 'On Trip',
  In_Shop: 'In Shop',
  Retired: 'Retired',
} as const;

export interface Vehicle {
  id: number;
  registrationNumber: string;
  name: string;
  type: string;
  maxLoadCapacity: number;
  odometer: number;
  acquisitionCost: number;
  status: VehicleStatus;
  region: string;
  createdAt: string;
  updatedAt: string;
}

export interface VehicleInput {
  /** @minLength 1 */
  registrationNumber: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  type: string;
  /** @minimum 0 */
  maxLoadCapacity: number;
  /** @minimum 0 */
  odometer?: number;
  /** @minimum 0 */
  acquisitionCost: number;
  /** @minLength 1 */
  region: string;
}

export interface VehicleUpdate {
  /** @minLength 1 */
  registrationNumber?: string;
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  type?: string;
  /** @minimum 0 */
  maxLoadCapacity?: number;
  /** @minimum 0 */
  odometer?: number;
  /** @minimum 0 */
  acquisitionCost?: number;
  /** @minLength 1 */
  region?: string;
  status?: VehicleStatus;
}

export type DriverStatus = typeof DriverStatus[keyof typeof DriverStatus];


export const DriverStatus = {
  Available: 'Available',
  On_Trip: 'On Trip',
  Off_Duty: 'Off Duty',
  Suspended: 'Suspended',
} as const;

export interface Driver {
  id: number;
  name: string;
  licenseNumber: string;
  licenseCategory: string;
  licenseExpiryDate: string;
  contactNumber: string;
  safetyScore: number;
  status: DriverStatus;
  createdAt: string;
  updatedAt: string;
}

export interface DriverInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  licenseNumber: string;
  /** @minLength 1 */
  licenseCategory: string;
  licenseExpiryDate: string;
  /** @minLength 1 */
  contactNumber: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  safetyScore?: number;
}

export interface DriverUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  licenseNumber?: string;
  /** @minLength 1 */
  licenseCategory?: string;
  licenseExpiryDate?: string;
  /** @minLength 1 */
  contactNumber?: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  safetyScore?: number;
  status?: DriverStatus;
}

export type TripStatus = typeof TripStatus[keyof typeof TripStatus];


export const TripStatus = {
  Draft: 'Draft',
  Dispatched: 'Dispatched',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;

export interface Trip {
  id: number;
  source: string;
  destination: string;
  vehicleId: number;
  driverId: number;
  cargoWeight: number;
  plannedDistance: number;
  status: TripStatus;
  /** @nullable */
  actualDistance: number | null;
  /** @nullable */
  fuelConsumed: number | null;
  createdAt: string;
  /** @nullable */
  dispatchedAt: string | null;
  /** @nullable */
  completedAt: string | null;
  /** @nullable */
  cancelledAt: string | null;
}

export interface TripInput {
  /** @minLength 1 */
  source: string;
  /** @minLength 1 */
  destination: string;
  vehicleId: number;
  driverId: number;
  /** @minimum 0 */
  cargoWeight: number;
  /** @minimum 0 */
  plannedDistance: number;
}

export interface TripUpdate {
  /** @minLength 1 */
  source?: string;
  /** @minLength 1 */
  destination?: string;
  vehicleId?: number;
  driverId?: number;
  /** @minimum 0 */
  cargoWeight?: number;
  /** @minimum 0 */
  plannedDistance?: number;
}

export interface TripCompleteInput {
  /** @minimum 0 */
  finalOdometer: number;
  /** @minimum 0 */
  fuelConsumed: number;
}

export type MaintenanceStatus = typeof MaintenanceStatus[keyof typeof MaintenanceStatus];


export const MaintenanceStatus = {
  Open: 'Open',
  Closed: 'Closed',
} as const;

export interface MaintenanceLog {
  id: number;
  vehicleId: number;
  description: string;
  cost: number;
  status: MaintenanceStatus;
  startedAt: string;
  /** @nullable */
  closedAt: string | null;
  createdAt: string;
}

export interface MaintenanceLogInput {
  vehicleId: number;
  /** @minLength 1 */
  description: string;
  /** @minimum 0 */
  cost: number;
}

export interface MaintenanceLogUpdate {
  /** @minLength 1 */
  description?: string;
  /** @minimum 0 */
  cost?: number;
}

export interface FuelLog {
  id: number;
  vehicleId: number;
  /** @nullable */
  tripId?: number | null;
  liters: number;
  cost: number;
  date: string;
  createdAt: string;
}

export interface FuelLogInput {
  vehicleId: number;
  tripId?: number;
  /** @minimum 0 */
  liters: number;
  /** @minimum 0 */
  cost: number;
  date: string;
}

export type ExpenseCategory = typeof ExpenseCategory[keyof typeof ExpenseCategory];


export const ExpenseCategory = {
  Toll: 'Toll',
  Insurance: 'Insurance',
  Permit: 'Permit',
  Other: 'Other',
} as const;

export interface Expense {
  id: number;
  vehicleId: number;
  category: ExpenseCategory;
  amount: number;
  date: string;
  description: string;
  createdAt: string;
}

export interface ExpenseInput {
  vehicleId: number;
  category: ExpenseCategory;
  /** @minimum 0 */
  amount: number;
  date: string;
  /** @minLength 1 */
  description: string;
}

export interface DashboardSummary {
  activeVehicles: number;
  availableVehicles: number;
  vehiclesInMaintenance: number;
  retiredVehicles: number;
  activeTrips: number;
  pendingTrips: number;
  driversOnDuty: number;
  fleetUtilization: number;
  expiringLicenses: number;
}

export interface ActivityEntry {
  id: string;
  type: string;
  message: string;
  timestamp: string;
}

export interface VehicleCostReportRow {
  vehicleId: number;
  registrationNumber: string;
  name: string;
  totalDistance: number;
  totalFuelLiters: number;
  /** @nullable */
  fuelEfficiency: number | null;
  fuelCost: number;
  maintenanceCost: number;
  expenseCost: number;
  totalOperationalCost: number;
  revenue: number;
  /** @nullable */
  roi: number | null;
}

export type ListVehiclesParams = {
type?: string;
status?: VehicleStatus;
region?: string;
search?: string;
};

export type ListDriversParams = {
status?: DriverStatus;
search?: string;
};

export type ListTripsParams = {
status?: TripStatus;
};

export type ListMaintenanceLogsParams = {
vehicleId?: number;
status?: MaintenanceStatus;
};

export type ListFuelLogsParams = {
vehicleId?: number;
};

export type ListExpensesParams = {
vehicleId?: number;
category?: ExpenseCategory;
};

export type GetDashboardSummaryParams = {
type?: string;
status?: string;
region?: string;
};