# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TransitOps fleet operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List drivers
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/DriverStatus"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [drivers]
      summary: Register a new driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDriver
      tags: [drivers]
      summary: Update a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverUpdate"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDriver
      tags: [drivers]
      summary: Delete a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /maintenance-logs:
    get:
      operationId: listMaintenanceLogs
      tags: [maintenance]
      summary: List maintenance logs
      parameters:
        - name: vehicleId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/MaintenanceStatus"
      responses:
        "200":
          description: List of maintenance logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceLog"
    post:
      operationId: createMaintenanceLog
      tags: [maintenance]
      summary: Create a maintenance record (sets vehicle to In Shop)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceLog"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /maintenance-logs/{id}:
    patch:
      operationId: updateMaintenanceLog
      tags: [maintenance]
      summary: Update a maintenance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceLogUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceLog"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /maintenance-logs/{id}/close:
    post:
      operationId: closeMaintenanceLog
      tags: [maintenance]
      summary: Close a maintenance record (restores vehicle to Available unless retired)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceLog"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /fuel-logs:
    get:
      operationId: listFuelLogs
      tags: [fuel]
      summary: List fuel logs
      parameters:
        - name: vehicleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of fuel logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FuelLog"
    post:
      operationId: createFuelLog
      tags: [fuel]
      summary: Record a fuel log
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FuelLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FuelLog"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /fuel-logs/{id}:
    delete:
      operationId: deleteFuelLog
      tags: [fuel]
      summary: Delete a fuel log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: vehicleId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/ExpenseCategory"
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Record an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{id}:
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    DriverStatus:
      type: string
      enum: [Available, On Trip, Off Duty, Suspended]
    Driver:
      type: object
      required:
        [
          id,
          name,
          licenseNumber,
          licenseCategory,
          licenseExpiryDate,
          contactNumber,
          safetyScore,
          status,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        name: { type: string }
        licenseNumber: { type: string }
        licenseCategory: { type: string }
        licenseExpiryDate: { type: string, format: date }
        contactNumber: { type: string }
        safetyScore: { type: number }
        status:
          $ref: "#/components/schemas/DriverStatus"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DriverInput:
      type: object
      required:
        [name, licenseNumber, licenseCategory, licenseExpiryDate, contactNumber]
      properties:
        name: { type: string, minLength: 1 }
        licenseNumber: { type: string, minLength: 1 }
        licenseCategory: { type: string, minLength: 1 }
        licenseExpiryDate: { type: string, format: date }
        contactNumber: { type: string, minLength: 1 }
        safetyScore: { type: number, minimum: 0, maximum: 100 }
    DriverUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        licenseNumber: { type: string, minLength: 1 }
        licenseCategory: { type: string, minLength: 1 }
        licenseExpiryDate: { type: string, format: date }
        contactNumber: { type: string, minLength: 1 }
        safetyScore: { type: number, minimum: 0, maximum: 100 }
        status:
          $ref: "#/components/schemas/DriverStatus"
...
    MaintenanceStatus:
      type: string
      enum: [Open, Closed]
    MaintenanceLog:
      type: object
      required:
        [
          id,
          vehicleId,
          description,
          cost,
          status,
          startedAt,
          closedAt,
          createdAt,
        ]
      properties:
        id: { type: integer }
        vehicleId: { type: integer }
        description: { type: string }
        cost: { type: number }
        status:
          $ref: "#/components/schemas/MaintenanceStatus"
        startedAt: { type: string, format: date-time }
        closedAt:
          type: ["string", "null"]
          format: date-time
        createdAt: { type: string, format: date-time }
    MaintenanceLogInput:
      type: object
      required: [vehicleId, description, cost]
      properties:
        vehicleId: { type: integer }
        description: { type: string, minLength: 1 }
        cost: { type: number, minimum: 0 }
    MaintenanceLogUpdate:
      type: object
      properties:
        description: { type: string, minLength: 1 }
        cost: { type: number, minimum: 0 }
    FuelLog:
      type: object
      required: [id, vehicleId, liters, cost, date, createdAt]
      properties:
        id: { type: integer }
        vehicleId: { type: integer }
        tripId:
          type: ["integer", "null"]
        liters: { type: number }
        cost: { type: number }
        date: { type: string, format: date }
        createdAt: { type: string, format: date-time }
    FuelLogInput:
      type: object
      required: [vehicleId, liters, cost, date]
      properties:
        vehicleId: { type: integer }
        tripId: { type: integer }
        liters: { type: number, minimum: 0 }
        cost: { type: number, minimum: 0 }
        date: { type: string, format: date }
    ExpenseCategory:
      type: string
      enum: [Toll, Insurance, Permit, Other]
    Expense:
      type: object
      required:
        [id, vehicleId, category, amount, date, description, createdAt]
      properties:
        id: { type: integer }
        vehicleId: { type: integer }
        category:
          $ref: "#/components/schemas/ExpenseCategory"
        amount: { type: number }
        date: { type: string, format: date }
        description: { type: string }
        createdAt: { type: string, format: date-time }
    ExpenseInput:
      type: object
      required: [vehicleId, category, amount, date, description]
      properties:
        vehicleId: { type: integer }
        category:
          $ref: "#/components/schemas/ExpenseCategory"
        amount: { type: number, minimum: 0 }
        date: { type: string, format: date }
        description: { type: string, minLength: 1 }
...