# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TransitOps fleet operations API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseCategory } from './expenseCategory';

export interface Expense {
  id: number;
  vehicleId: number;
  category: ExpenseCategory;
  amount: number;
  date: Date;
  description: string;
  createdAt: Date;
}