# lib/api-zod/src/generated/types/maintenanceLogInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TransitOps fleet operations API
 * OpenAPI spec version: 0.1.0
 */

export interface MaintenanceLogInput {
  vehicleId: number;
  /** @minLength 1 */
  description: string;
  /** @minimum 0 */
  cost: number;
}